wip - experiment with dynamic imports
